tests/test_deduplication.py
"""
Tests para deduplicación de eventos en extracted_events.
"""

import pytest
from datetime import datetime, timedelta


@pytest.mark.asyncio
async def test_dedup_logic():
    """
    Test de lógica de deduplicación.
    Verifica que eventos duplicados (mismo título, source, ventana de tiempo)
    sean detectados correctamente.
    """
    # Nota: El script dedup_extracted_events.py tiene la función dedup_extracted_events
    # Este test es un placeholder hasta que se refactorice el script
    import sys
    from pathlib import Path
    sys.path.insert(0, str(Path(__file__).parent.parent))
    
    # Mock de eventos duplicados
    events = [
        {
            "id": 1,
            "title": "Reunión Test",
            "source": "gmail",
            "start_at": "2025-12-16T10:00:00Z",
            "status": "pending",
        },
        {
            "id": 2,
            "title": "Reunión Test",
            "source": "gmail",
            "start_at": "2025-12-16T10:15:00Z",  # 15 min después, dentro de ventana 90 min
            "status": "pending",
        },
        {
            "id": 3,
            "title": "Reunión Test",
            "source": "calendly",
            "start_at": "2025-12-16T10:00:00Z",  # Mismo tiempo pero diferente source
            "status": "pending",
        },
    ]
    
    # La función deduplicate_events debería identificar que id=1 e id=2 son duplicados
    # (mismo título, mismo source, dentro de 90 min)
    # Pero id=3 no es duplicado (diferente source)
    
    # Nota: Este test requiere que el script dedup_extracted_events.py
    # esté implementado y accesible. Por ahora es un placeholder.
    assert True  # Placeholder hasta implementar lógica de dedup


@pytest.mark.asyncio
async def test_extracted_events_source_field():
    """
    Test que verifica que extracted_events tiene el campo 'source'.
    """
    # Este test verifica que la estructura de datos incluye 'source'
    # Requiere conexión a Supabase para ejecutarse completamente
    assert True  # Placeholder

